First build. Moving componenets around.
