+++ src/components/setting/following.tsx
@@ -1,8 +1,8 @@
 import { Box, Heading } from "@chakra-ui/react"
 import React, { useContext } from "react"
-import { UserContext } from "../../../context/userContext"
+import { UserContext } from "../../context/userContext"
 
-const Theme = (): JSX.Element => {
+const Following = (): JSX.Element => {
   // User Profile Context
   const { userProfile } = useContext(UserContext)
   // Conditionally render the skeleton loading effects
@@ -10,9 +10,9 @@ const Theme = (): JSX.Element => {
 
   return (
     <Box>
-      <Heading>{userProfile.name}'s Theme</Heading>
+      <Heading>{userProfile.name}'s Following</Heading>
     </Box>
   )
 }
 
-export default Theme;
+export default Following;
