+++ src/components/setting/theme.tsx
@@ -1,8 +1,8 @@
 import { Box, Heading } from "@chakra-ui/react"
 import React, { useContext } from "react"
-import { UserContext } from "../../../context/userContext"
+import { UserContext } from "../../context/userContext"
 
-const Accessibility = (): JSX.Element => {
+const Theme = (): JSX.Element => {
   // User Profile Context
   const { userProfile } = useContext(UserContext)
   // Conditionally render the skeleton loading effects
@@ -10,9 +10,9 @@ const Accessibility = (): JSX.Element => {
 
   return (
     <Box>
-      <Heading>{userProfile.name}'s Accessibility</Heading>
+      <Heading>{userProfile.name}'s Theme</Heading>
     </Box>
   )
 }
 
-export default Accessibility;
+export default Theme;
