+++ src/pages/search/index.tsx
@@ -14,7 +14,7 @@ import {
 } from "@chakra-ui/react";
 import React, { Fragment, useEffect, useRef, useState } from "react";
 import { FaCaretDown } from "react-icons/fa";
-import SearchResults from "./searchResults";
+import SearchResults from "../../components/search/searchResults";
 import { useGetAllProfilesQuery, useProjectsQuery } from "../../generated/graphql";
 import { withUrqlClient } from "next-urql";
 import { Profile } from '../../../server/src/db/entities/Profile'
